Poingr.ContactMechanics

Poingr.ContactMechanics.Contact

Poingr.ContactMechanics.Contact — Type

Contact(:sticky)
Contact(:slip; sep)
Contact(:friction, coef; sep)

Contact condition handling contact mechanics in MPM. Following conditions are available:

:sticky: Continuum body is sticked on the boundary surface.

:slip: Continuum body is slipped on the boundary surface.

:friction: Continuum body is slipped with frictional coefficient coef on the boundary surface.

If sep is true, continuum body can leave from the boundary surface.

Examples

julia> Contact(:sticky)
Contact(:sticky)

julia> Contact(:slip, sep = true)
Contact(:slip; sep = true)

julia> Contact(:friction, 0.3, sep = true)
Contact(:friction, 0.3; sep = true)

(::Contact)(v::Vec, n::Vec)

Compute velocity v caused by contact. The other quantities, which are equivalent to velocity such as momentum and force, are also available. n is the normal unit vector.

Examples

julia> contact = Contact(:slip, sep = false);

julia> v = Vec(1.0, -1.0); n = Vec(0.0, -1.0);

julia> v + contact(v, n)
2-element Vec{2, Float64}:
 1.0
 0.0