Poingr.DofHelpers

Poingr.DofHelpers.DofMap

Poingr.DofHelpers.DofMap

Poingr.DofHelpers.filter

Poingr.DofHelpers.map

Poingr.DofHelpers.ndofs

Poingr.DofHelpers.ndofs

Poingr.DofHelpers.DofMap — Type

DofMap(dims...)

Create function-like object for dof mapping. DofMap also behave like a bool array to activate/deactivate indices. To finalize activations and perform numbering dofs, use count!(::DofMap).

count!(::DofMap)

(DofMap)(index...; [dof])

DofHelpers.map(::DofMap, inds; [dof])

DofHelpers.filter(::DofMap, inds)

Poingr.DofHelpers.DofMap — Method

(::DofMap)(index...; [dof])

Return dof from given nodal index. Use DofHelpers.map(::DofMap, indices::AbstractArray) for multiple indices.

Examples

julia> dofmap = DofMap(5, 5)
5×5 DofMap{2}:
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

julia> dofmap[1:2, 2:3] .= true; dofmap
5×5 DofMap{2}:
 0  1  1  0  0
 0  1  1  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

julia> count!(dofmap)
4

julia> dofmap(1, 1) === nothing
true

julia> dofmap(2, 2)
2

julia> dofmap(2, 2, dof = 1)
2:2

julia> dofmap(2, 2, dof = 2)
3:4

Base.count! — Method

count!(::DofMap)

Count active indices and numbering dofs. Returned integer is the number of active indices.

Examples

julia> dofmap = DofMap(5, 5)
5×5 DofMap{2}:
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

julia> dofmap[1:2, 2:3] .= true; dofmap
5×5 DofMap{2}:
 0  1  1  0  0
 0  1  1  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

julia> count!(dofmap)
4

Poingr.DofHelpers.filter — Method

DofHelpers.filter(::DofMap, inds::AbstractArray)
DofHelpers.filter!(::DofMap, output::Vector, inds::AbstractArray)

This is the same as inds[dofmap[inds]] but more effective.

Examples

julia> dofmap = DofMap(5, 5)
5×5 DofMap{2}:
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

julia> dofmap[1:2, 2:3] .= true; dofmap
5×5 DofMap{2}:
 0  1  1  0  0
 0  1  1  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

julia> count!(dofmap)
4

julia> DofHelpers.filter(dofmap, CartesianIndices((1:2, 1:2)))
2-element Vector{CartesianIndex{2}}:
 CartesianIndex(1, 2)
 CartesianIndex(2, 2)

Poingr.DofHelpers.map — Method

DofHelpers.map(::DofMap, indices::AbstractArray; [dof])
DofHelpers.map!(::DofMap, dofs::Vector{Int}, indices::AbstractArray; [dof])

Map nodal indices to dof indices. This is almost the same behavior as performing dofmap(index) at each elemnt of indices but nothing is skipped.

Examples

julia> dofmap = DofMap(5, 5)
5×5 DofMap{2}:
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

julia> dofmap[1:2, 2:3] .= true; dofmap
5×5 DofMap{2}:
 0  1  1  0  0
 0  1  1  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

julia> count!(dofmap)
4

julia> DofHelpers.map(dofmap, CartesianIndices((1:2, 1:2)); dof = 2)
4-element Vector{Int64}:
 1
 2
 3
 4

julia> DofHelpers.map(dofmap, [CartesianIndex(1, 2), CartesianIndex(2, 3)]; dof = 2)
4-element Vector{Int64}:
 1
 2
 7
 8

Poingr.DofHelpers.ndofs — Method

ndofs(::DofMap; [dof])

Return total number of dofs.