Poingr.Loggers

Poingr.Loggers.Logger

Poingr.Loggers.Logger — Type

Logger(logpoints::AbstractVector; progress = false)

Construct logger which handle with time event in the calculation. The workflow using Logger can be written as follows:

function workflow()
    logpoints = 0:0.5:5
    logger = Logger(logpoints)
    t = 0.0
    dt = 0.2
    timestamps = Float64[]

    while !isfinised(logger, t)
        #
        # 1. Calculations...
        #

        # 2. Update time step and logger
        update!(logger, t += dt)

        #
        # 3. Save data at log point
        #
        if islogpoint(logger)

            # linear index is available for numbering data
            i = logindex(logger)

            push!(timestamps, t)
        end
    end

    timestamps
end

workflow()

# output

11-element Vector{Float64}:
 0.2
 0.6000000000000001
 1.0
 1.5999999999999999
 2.1999999999999997
 2.6
 3.0000000000000004
 3.600000000000001
 4.000000000000001
 4.600000000000001
 5.000000000000002

As shown above example, islogpoint(logger) is true when time t become first greater than or equalt to logpoints.