Shape functions

Poingr.BSpline

Poingr.BSplinePosition

Poingr.ShapeValues

Poingr.update!

Poingr.ShapeValues — Method

Poingr.ShapeValues(::ShapeFunction)
Poingr.ShapeValues(::Type{T}, ::ShapeFunction)

Construct object storing value of ShapeFunction.

Examples

julia> sv = Poingr.ShapeValues(QuadraticBSpline{2}());

julia> update!(sv, Grid(0:3, 0:3), Vec(1, 1));

julia> sum(sv.N)
1.0

julia> sum(sv.∇N)
2-element Vec{2, Float64}:
 5.551115123125783e-17
 5.551115123125783e-17

Poingr.update! — Function

update!(::ShapeValues, grid::Grid, x::Vec)
update!(::ShapeValues, grid::Grid, indices::AbstractArray, x::Vec)

Update value of shape function at x with each grid node.

Examples

julia> sv = Poingr.ShapeValues(QuadraticBSpline{2}());

julia> update!(sv, Grid(0:3, 0:3), Vec(1, 1));

julia> sum(sv.N)
1.0

julia> update!(sv, Grid(0:3, 0:3), Vec(1, 1), CartesianIndices((1:2, 1:2)));

julia> sum(sv.N)
0.765625

Poingr.BSpline — Type

BSpline{order, dim}()
LinearBSpline{dim}()
QuadraticBSpline{dim}()
CubicBSpline{dim}()

Create B-spline shape function.

Examples

julia> f = LinearBSpline{2}()
LinearBSpline{2}()

julia> f(Vec(0.5, 0.5))
0.25

Poingr.BSplinePosition — Type

Poingr.BSplinePosition(; nth::Int, dir::Int)

Position of node at ξ = 0. The node is located at nth away from bound with dir direction.

Examples

julia> Poingr.BSplinePosition([1,2,3,4,5], 2)
BSplinePosition(nth = 1, dir = 1)

julia> Poingr.BSplinePosition([1,2,3,4,5], 4)
BSplinePosition(nth = 1, dir = -1)

julia> pos = Poingr.BSplinePosition([1,2,3,4,5], 1)
BSplinePosition(nth = 0, dir = 0)

julia> f = QuadraticBSpline{0}()
QuadraticBSpline{0}()

julia> f(0.0, pos)
1.0