Contact mechanics

Marble.CoulombFriction

Marble.CoulombFriction

Marble.contacted

Marble.CoulombFriction — Method

CoulombFriction(:sticky)

This is the same as the CoulombFriction(; μ = Inf).

CoulombFriction(:slip; separation = false)

This is the same as the CoulombFriction(; μ = 0, separation = false).

CoulombFriction(:free)

No boundary condition. This is the same as the CoulombFriction(; μ = NaN).

Marble.CoulombFriction — Method

CoulombFriction(; parameters...)

Frictional contact using Mohr-Coulomb criterion.

Parameters

μ: friction coefficient (use μ or ϕ)

ϕ: friction angle (radian)

c: cohesion (default: 0)

separation: true or false (default: false).

If separation is true, continuum body can leave from the boundary surface.

Marble.contacted — Method

contacted(::CoulombFriction, q::Vec, n::Vec)

Compute q caused by contact. q can be used as velocity (relative velocity) v or force f. n is the unit vector normal to the surface.

Examples

julia> cond = CoulombFriction(:slip, separation = false);

julia> v = Vec(1.0, -1.0); n = Vec(0.0, 1.0);

julia> v + contacted(cond, v, n)
2-element Vec{2, Float64}:
 1.0
 0.0

julia> v + contacted(CoulombFriction(:free), v, n) == v
true